requirements.txt
requests
PyYAML
python-dotenv
discord.py